+++ Discordbot/supabase/functions/reminder/index.js
@@ -1,0 +1,1 @@
+"use strict";
